- Just getting some rough functioning code in that at least binds events to the listbox
