+++ WaywardHorizonsWPF/WaywardHorizons/Environment/Lake.xaml
@@ -21,10 +21,12 @@
             <TextBlock.Effect>
                 <DropShadowEffect/>
             </TextBlock.Effect><Run Text="Health:"/><LineBreak/><Run Text="Supplies:"/></TextBlock>
-        <TextBlock Uid="tbEvents" HorizontalAlignment="Left" Height="206" Margin="33,105,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="367" FontWeight="Bold" FontSize="14" Foreground="White">
+        <TextBlock Uid="tbEvents" HorizontalAlignment="Left" Margin="36,287,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="367" FontWeight="Bold" FontSize="14" Foreground="White">
             <TextBlock.Effect>
                 <DropShadowEffect/>
-            </TextBlock.Effect><Run Text="Event Text Here..."/><LineBreak/><Run/><LineBreak/><Run/><LineBreak/><Run/><LineBreak/><Run/><LineBreak/><Run/><LineBreak/><Run/><LineBreak/><Run/><LineBreak/><Run/><LineBreak/><Run Text="Enter Choice Below..."/></TextBlock>
+            </TextBlock.Effect><Run Text="Enter Choice Below..."/>
+        </TextBlock>
+        <ListBox ItemsSource="{Binding LocationEvents}" x:Name="lbEvents" Margin="33,80,395,178" Background="{x:Null}" />
         <TextBox HorizontalAlignment="Left" Margin="33,311,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="143" Height="34"/>
         <Button x:Name="Submit" Content="Submit" HorizontalAlignment="Left" Margin="192,311,0,0" VerticalAlignment="Top" Height="34" Width="55" Click="Button_Click"/>
